chore: makes each class have there own file
