+++ FrankenToilet/greycsont/DestoryTimer.cs
@@ -1,0 +1,18 @@
+using UnityEngine;
+
+namespace FrankenToilet.greycsont;
+
+public class DestoryTimer : MonoBehaviour
+{
+    public float lifetime;
+    private float timer = 0f;
+
+    void Update()
+    {
+        timer += Time.unscaledDeltaTime;
+        if (timer >= lifetime)
+        {
+            Destroy(gameObject);
+        }
+    }
+}
