+++ FrankenToilet/greycsont/Direction.cs
@@ -1,0 +1,10 @@
+
+namespace FrankenToilet.greycsont;
+
+public enum Direction
+{
+    Backwards = 0,
+    Left = 1,
+    Upwards = 2,
+    Right = 3
+}
