+++ FrankenToilet/greycsont/Patches/ShotgunHammerPatches.cs
@@ -13,6 +13,8 @@ namespace FrankenToilet.greycsont;
 [HarmonyPatch(typeof(ShotgunHammer))]
 public static class ShotgunHammerPatch
 {
+    public static ShotgunHammer lastActiveHammer;
+
     private static readonly MethodInfo negate = AccessTools.Method(typeof(Vector3), "op_UnaryNegation");
 
     private static readonly MethodInfo random4 = AccessTools.Method(typeof(DirectionRandomizer), nameof(DirectionRandomizer.Randomize4Dir));
@@ -29,7 +31,7 @@ public static class ShotgunHammerPatch
     [HarmonyPatch(nameof(ShotgunHammer.Impact))]
     public static void ImpactPatch(ShotgunHammer __instance)
     {
-        HammerTracker.lastActiveHammer = __instance;
+        lastActiveHammer = __instance;
         DirectionRandomizer.GenerateRandomDirection();
     }
 
@@ -48,7 +50,7 @@ public static class ShotgunHammerPatch
         var num6 = matcher.Instruction;
 
         matcher.InsertAndAdvance(
-            new CodeInstruction(num6.opcode, num6.operand), // 再次加载 num6
+            new CodeInstruction(num6.opcode, num6.operand),
             new CodeInstruction(OpCodes.Call, GenerateArrowImage)
         );
 
@@ -74,8 +76,4 @@ public static class ShotgunHammerPatch
 }
 
 
-public static class HammerTracker
-{
-    public static ShotgunHammer lastActiveHammer;
-}
